Make changes that (hopefully) impact overall perf
Moved source maps back inline, ready for istanbul v1.

The main changes for perf are trying to use the project version and file versions to decide when to increment the project version. It’s doing the least amount of fs hits possible, but we should still be able to optimize the start up use-case since that hits things double-time. First, TypeScript, and second, node.