Fix bug where js extension gets not registered when lazy is set to true and allowJS is used via the configuration
This bugfix solves a problem when using ts-node/register with a not ts file as entry point.
The lazy option prevents the evaluation of the tsconfig.json file until the first ts file is loaded.
In scenarios where only js files are used and ts-node is used as transpiler this will never happen.
The result is that ts-node is never used for javascript files even allowJS is defined in the tsconfig.json.

This pull request solves the issue by reading the configuration as soon as possible to determine if `allowJS` is set
to true or not. If so the js extension get registered.