Stop keeping REPL state after successful commands
Right now, the REPL keeps the input between successful commands, leading
to some unexpected behaviours like the following:

```
~/src  ts-node
> 5
5
> +3
8
> -2
1
```

Instead of treating each line as a separate unary numeric expression,
ts-node treats the first one correctly, but then groups the 1st and
second to get "5 +3" = 8, and then the 2nd and 3rd to get "+3 -2" = 1.

This commit force-resets the input state after a successful command to
prevent this from happening.

I couldn't manage to reproduce this in tests since AFAICT it only
happens when using the REPL with a TTY, and I couldn't get this
environment to work in tests. Very open to trying again if there are
alternatives I missed.