Forward whole context to ESM hooks
Ensure that all properties of the context objects passed to the
`resolve` and `load` ESM hooks are forwarded to the default hooks.

This change is necessary to for using [import assertions in Node
v17][1]. Node v17 requires these import assertions for JSON modules. If
the assertions are not passed down to the default `load` function, Node
refuses to import the JSON module.

To prevent similar issues we also forward the context for the `resolve`
hook.

[1]: https://nodejs.org/dist/latest-v17.x/docs/api/esm.html#import-assertions
Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>