loader: TSError that survives thread comms channel
When an error is thrown in the loader thread, it must be passed through
the comms channel to be printed in the main thread.

Node has some heuristics to try to reconstitute errors properly, but
they don't function very well if the error has a custom inspect method,
or properties that are not compatible with JSON.stringify, so the
TSErrors raised by the source transforms don't get printed in any sort
of useful way.

This catches those errors, and creates a new error that can go through
the comms channel intact.

Another possible approach would be to update the shape of the errors
raised by source transforms, but that would be a much more extensive
change with further reaching consequences.