loader: Set default 'pretty' option properly
Set the default `options.pretty` value based on stderr rather than
stdout, as this is where errors are printed.

The loader thread does not get a process.stderr.isTTY set, because its
"stderr" is actually a pipe. If `options.pretty` is not set explicitly,
the GlobalPreload's `context.port` is used to send a message from the
main thread indicating the state of stderr.isTTY.

Adds `Service.setPrettyErrors` method to enable setting this value when
needed.