use built-in source map support in node v20
The @cspotcode/source-map-support module does not function properly on
Node 20, resulting in incorrect stack traces. Fortunately, the built-in
source map support in Node is now quite reliable. This does have the
following (somewhat subtle) changes to error output:

- When a call site is in a method defined within a constructor function,
  it picks up the function name *as well as* the type name and method
  name. So, in tests where a method is called and throws within the a
  function constructor, we see `Foo.Foo.bar` instead of `Foo.bar`.
- Call sites displays show filenames instead of file URLs.
- The call site display puts the `^` character under the `throw` rather
  than the construction point of the error object. This is closer to how
  normal un-transpiled JavaScript behaves, and thus somewhat
  preferrable, but isn't possible when all we have to go on is the Error
  stack property, so it is a change.

I haven't been able to figure out why exactly, but the call sites appear
to be somewhat different in the repl/eval contexts as a result of this
change. It almost seems like the @cspotcode/source-map-support was
applying source maps to the vm-evaluated scripts, but I don't see how
that could be, and in fact, there's a comment in the code stating that
that *isn't* the case. But the line number showing up in an Error.stack
property is `1` prior to this change (matching the location in the TS
source) and is `2` afterwards (matching the location in the compiled
JS).

An argument could be made that specific line numbers are a bit
meaningless in a REPL anyway, and the best approach is to just make
those tests accept either result. One possible approach to provide
built-in source map support for the repl would be to refactor the
`appendCompileAndEvalInput` to diff and append the *input* TS, and
compile within the `runInContext` method. If the transpiled code was
prepended with `process.setSourceMapsEnabled(true);`, then Error stacks
and call sites would be properly source mapped by Node internally.